dom/animation/test/chrome/test_keyframe_effect_xrays.html

<!doctype html>
<head>
<meta charset=utf-8>
<script type="application/javascript" src="../testharness.js"></script>
<script type="application/javascript" src="../testharnessreport.js"></script>
<script type="application/javascript" src="../testcommon.js"></script>
</head>
<body>
<a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1414674"
  target="_blank">Mozilla Bug 1414674</a>
<div id="log"></div>
<iframe id="iframe"
  src="http://example.org/tests/dom/animation/test/chrome/file_animate_xrays.html"></iframe>
<script>
'use strict';

var win = document.getElementById('iframe').contentWindow;

async_test(function(t) {
  window.addEventListener('load', t.step_func(function() {
    var target = win.document.getElementById('target');
    var effect = new win.KeyframeEffect(target, [
      {opacity: 1, offset: 0},
      {opacity: 0, offset: 1},
    ], {duration: 100 * MS_PER_SEC, fill: "forwards"});
    // The frames object should be accessible via x-ray.
    var frames = effect.getKeyframes();
    assert_equals(frames.length, 2,
                  "frames for KeyframeEffect ctor should be non-zero");
    assert_equals(frames[0].opacity, "1",
                  "first frame opacity for KeyframeEffect ctor should be specified value");
    assert_equals(frames[0].computedOffset, 0,
                  "first frame offset for KeyframeEffect ctor should be 0");
    assert_equals(frames[1].opacity, "0",
                  "last frame opacity for KeyframeEffect ctor should be specified value");
    assert_equals(frames[1].computedOffset, 1,
                  "last frame offset for KeyframeEffect ctor should be 1");
    var animation = new win.Animation(effect, document.timeline);
    animation.play();
    t.done();
  }));
}, 'Calling KeyframeEffect() ctor across x-rays');

</script>
</body>
